Fix division by 0 error and use a faster FFT
Every 0 entry in the filter bank and the power spectrum is now set to
0.000003. The number is arbitrary and may need refinement, but at least
it won't try to do log_10(0). This fixes #31.

It now computes a 512-point FFT because 512 is a power of 2 and
therefore expedites computation.

For real input (which we are using) the FFT contains redundant
information. numpy.fft.rfft avoids the redundant calculation,
making it run significantly faster.